+++ src/nationalarchives/components/checkboxes/analytics.js
@@ -1,0 +1,25 @@
+import { valueGetters } from "../../lib/analytics-helpers.mjs";
+
+export default [
+  {
+    scope: ".tna-checkboxes",
+    areaName: "checkboxes",
+    events: [
+      {
+        eventName: "toggle",
+        targetElement: ".tna-checkboxes__item input",
+        on: "change",
+        data: {
+          state: valueGetters.checked,
+          // eslint-disable-next-line no-unused-vars
+          value: ($el, $scope, event) => $el.parentNode.innerText,
+          // eslint-disable-next-line no-unused-vars
+          group: ($el, $scope, event) =>
+            $scope
+              .closest(".tna-form__group")
+              ?.querySelector(".tna-form__heading")?.innerText,
+        },
+      },
+    ],
+  },
+];