+++ tasks/generate-fixture-html.js
@@ -6,9 +6,10 @@ const { renderNunjucks } = require("./lib/nunjucks");
 
 const componentsDirectory = "src/nationalarchives/components/";
 const componentFixturesFile = "/fixtures.json";
+const fixturesOutputDirectory = "fixtures-html";
 
-if (!fs.existsSync("temp")) {
-  fs.mkdirSync("temp");
+if (!fs.existsSync(fixturesOutputDirectory)) {
+  fs.mkdirSync(fixturesOutputDirectory);
 }
 
 const components = globSync(
@@ -29,7 +30,7 @@ components.forEach((component) => {
   componentFixtures.fixtures.forEach((fixture) => {
     const result = renderNunjucks(componentNunjucks, fixture.options);
     fs.writeFile(
-      `temp/${component}-${fixture.name
+      `${fixturesOutputDirectory}/${component}-${fixture.name
         .replace(/[^0-9a-z]/gi, "-")
         .toLowerCase()}.html`,
       result,